backend/services/restaurant-worker-service/requirements.txt
kafka-python==2.0.2
SQLAlchemy==2.0.48
pymysql==1.1.0
python-dotenv==1.2.2
pymongo==4.10.1
